requirements.txt
streamlit
PyPDF2
python-docx
pillow
pytesseract
pdf2image
langdetect
transformers
torch==2.0.1
sentencepiece
diff-match-patch
